+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Changelog
 
-## [Unreleased]
+## [Unchecked]
 
 ### Fixed
 
 - Crash when using `react-native-reanimated` along with `bitmovin-player-react-native` and playing a live stream
+- Expo config plugin feature configurations taking no effect
 
 ## [1.0.0] - 2025-08-04